+++ src/Modal/Modal.tsx
@@ -1,6 +1,17 @@
-import React, { FC, ReactNode, PropsWithChildren } from "react";
+import React, { type FC, type ReactNode, type PropsWithChildren } from "react";
 import Transition from "react-transition-group/Transition";
-import ReactModal, { Props as ReactModalProps } from "react-modal";
+
+import {
+  FloatingFocusManager,
+  FloatingPortal,
+  FloatingOverlay,
+  useFloating,
+  useClick,
+  useDismiss,
+  useRole,
+  useInteractions,
+  type ReferenceType,
+} from "@floating-ui/react-dom-interactions";
 
 import { ConditionalWrapper, DisplayIf as If } from "src/Conditional";
 import { appendClass as ac } from "src/utils";
@@ -8,6 +19,7 @@ import { appendClass as ac } from "src/utils";
 import { ModalHeader } from "./Header";
 import { ModalBody } from "./Body";
 import { ModalFooter } from "./Footer";
+import { useFloatingContext } from "../FloatingProvider";
 
 /**
  * Modal
@@ -18,7 +30,7 @@ export type ModalSize = "small" | "default" | "large" | "full" | "fluid";
 export type ModalProps = PropsWithChildren<{
   size?: ModalSize;
   closeIcon?: boolean;
-  closeHandle?: (e) => void;
+  closeHandle?: () => void;
   title?: ReactNode;
   isOpen: boolean;
   autoClose?: boolean;
@@ -30,8 +42,12 @@ export type ModalProps = PropsWithChildren<{
   dialogProps?: React.ComponentProps<"div">;
   contentProps?: React.ComponentProps<"div">;
   maximize?: boolean;
-}> &
-  ReactModalProps;
+  refElement?: ReferenceType;
+  root?: Parameters<typeof FloatingPortal>[0]["root"];
+  lockScroll?: Parameters<typeof FloatingOverlay>[0]["lockScroll"];
+  ancestorScroll?: Parameters<typeof useDismiss>[1]["ancestorScroll"];
+}>;
+// & ReactModalProps;
 
 type ModalSizes = {
   Small: FC<ModalProps>;
@@ -60,18 +76,40 @@ export const Modal: ModalSizes & ModalComponents & FC<ModalProps> = ({
   maximize = false,
   children,
   isOpen,
-  ...props
+  refElement,
+  root: rootProvided,
+  lockScroll,
+  ancestorScroll,
 }) => {
   const [maximized, setMaximized] = React.useState(false);
+
   React.useEffect(() => setMaximized(false), [isOpen]);
   const realSize = React.useMemo(
     () => (maximized ? "full" : size),
     [maximized, size]
   );
 
-  const maximizeCb = React.useCallback(() => {
-    setMaximized((curr) => !curr);
-  }, []);
+  const modalContext = useFloatingContext();
+  const root: HTMLElement | null | undefined =
+    rootProvided ?? modalContext ? modalContext.rootRef.current : undefined;
+
+  const { reference, floating, context } = useFloating({
+    open: isOpen,
+    onOpenChange: (state) => (!state ? void closeHandle() : void 0),
+  });
+
+  React.useEffect(() => {
+    if (refElement) reference(refElement);
+  }, [refElement]);
+
+  const click = useClick(context);
+  const role = useRole(context, { role: "dialog" });
+  const dismiss = useDismiss(context, {
+    enabled: autoClose,
+    ancestorScroll,
+  });
+
+  const { getFloatingProps } = useInteractions([click, role, dismiss]);
 
   const nodeRef = React.useRef(null);
 
@@ -85,67 +123,78 @@ export const Modal: ModalSizes & ModalComponents & FC<ModalProps> = ({
       {...transitionEvents}
     >
       {(state) => (
-        <ReactModal
-          {...props}
-          onRequestClose={autoClose && closeHandle ? closeHandle : undefined}
-          overlayClassName="modal-backdrop"
-          isOpen={["entering", "entered"].includes(state)}
-          className={`modal${ac(realSize, `modal--${realSize}`)}${ac(
-            left,
-            "modal--left"
-          )}`}
-          closeTimeoutMS={
-            typeof animationDuration === "object"
-              ? animationDuration.exit
-              : animationDuration
-          }
-          overlayRef={(n) => (nodeRef.current = n)}
-        >
-          <div
-            className="modal__dialog"
-            {...dialogProps}
-            onClick={(e) => e.stopPropagation()}
+        <FloatingPortal root={root}>
+          <FloatingOverlay
+            className={`modal-backdrop${ac(
+              state === "exiting",
+              "modal-backdrop--before-close"
+            )}`}
+            lockScroll={lockScroll}
+            ref={nodeRef}
+            onClick={() => (autoClose ? closeHandle() : void 0)}
           >
-            <div className="modal__content" {...contentProps}>
-              <If condition={!!(closeIcon && closeHandle) || maximize}>
-                <ConditionalWrapper
-                  condition={!!(closeIcon && closeHandle) && maximize}
-                  wrapper={<div className="modal__close" />}
+            <FloatingFocusManager context={context}>
+              <div
+                ref={floating}
+                {...getFloatingProps({
+                  className: `modal${ac(realSize, `modal--${realSize}`)}${ac(
+                    left,
+                    "modal--left"
+                  )}`,
+                  onClick: () => (autoClose ? closeHandle() : void 0),
+                })}
+              >
+                <div
+                  className="modal__dialog"
+                  {...dialogProps}
+                  onClick={(e) => e.stopPropagation()}
                 >
-                  {Boolean(maximize) && (
-                    <a
-                      className={`${ac(
-                        !(closeIcon && closeHandle),
-                        "modal__close"
-                      )}${ac(closeIcon && closeHandle, "qtr-margin-right")}`}
-                      onClick={maximizeCb}
-                    >
-                      <span
-                        className={
-                          maximized ? "icon-minimize" : "icon-maximize"
-                        }
-                      />
-                    </a>
-                  )}
-                  {Boolean(closeIcon && closeHandle) && (
-                    <a
-                      className={!maximize ? "modal__close" : ""}
-                      onClick={closeHandle}
-                    >
-                      <span className="icon-close" />
-                    </a>
-                  )}
-                </ConditionalWrapper>
-              </If>
-              {Boolean(title) && (
-                <ModalHeader>
-                  <h1 className="modal__title">{title}</h1>
-                </ModalHeader>
-              )}
-              {children}
-            </div>
-          </div>
-        </ReactModal>
+                  <div className="modal__content" {...contentProps}>
+                    <If condition={!!(closeIcon && closeHandle) || maximize}>
+                      <ConditionalWrapper
+                        condition={!!(closeIcon && closeHandle) && maximize}
+                        wrapper={<div className="modal__close" />}
+                      >
+                        {Boolean(maximize) && (
+                          <a
+                            className={`${ac(
+                              !(closeIcon && closeHandle),
+                              "modal__close"
+                            )}${ac(
+                              closeIcon && closeHandle,
+                              "qtr-margin-right"
+                            )}`}
+                            onClick={() => setMaximized((curr) => !curr)}
+                          >
+                            <span
+                              className={
+                                maximized ? "icon-minimize" : "icon-maximize"
+                              }
+                            />
+                          </a>
+                        )}
+                        {Boolean(closeIcon && closeHandle) && (
+                          <a
+                            className={!maximize ? "modal__close" : ""}
+                            onClick={closeHandle}
+                          >
+                            <span className="icon-close" />
+                          </a>
+                        )}
+                      </ConditionalWrapper>
+                    </If>
+                    {Boolean(title) && (
+                      <ModalHeader>
+                        <h1 className="modal__title">{title}</h1>
+                      </ModalHeader>
+                    )}
+                    {children}
+                  </div>
+                </div>
+              </div>
+            </FloatingFocusManager>
+          </FloatingOverlay>
+        </FloatingPortal>
       )}
     </Transition>
   );