+++ src/Modal/PromptModal.tsx
@@ -20,7 +20,7 @@ export interface PromptModalProps<T extends React.ReactText> {
   title: ReactNode;
   question: ReactNode;
   onSave: (value: T) => void | Promise<void>;
-  onClose?: (e?) => void;
+  onClose?: () => void;
   initial?: T;
   type?: string;
   isOpen: boolean;